# Add a UserType flag for registered types, and use it to improve the performance for is_user_type.